Refactors fun2
Homework 4 exercise 1 asks you to refactor using wholemeal programming techniques. You are given a hint to use `takeWhile` and `iterate` to refactor `fun2`. I did this after looking at other solutions to see if there was any way to do this without `if`-`then`-`else`. I think the hangup for me is that although I see the logic here, I don't see this as an improvement over the original `fun2` and I think of refactoring as an exercise in improving code. I really dislike the where clause which just seems like a more awkward way of the original solution using guards. As opposed to the `fun1` refactor, I think the `fun2` refactor fails to make the code better.