Adds foldTree function
I get the gist of using `foldr` to convert a list into a binary tree (I used it to build the `MessageTree` in `LogAnalysis.hs`), but I had trouble implementing an `insert` function so that the tree always remains balanced. I looked at [Build balanced binary tree with foldr](http://stackoverflow.com/questions/16157203/build-balanced-binary-tree-with-foldr) on Stack Overflow for guidance. I'm not wild about all the `where` clauses, but it seems to work. I'll see if I can refactor it after a little more thought.